Add front ray casting

+++ src/grid.rs
@@ -3,19 +3,19 @@ use image::GenericImageView;
 
 pub struct Grid {
     pub tiles: Vec<u8>,
-    pub width: u32,
-    pub height: u32,
+    pub width: i32,
+    pub height: i32,
     pub tile_size: i32,
 }
 
 impl Grid {
     pub fn init(&mut self) {
         let grid_image = image::open("assets/grid.png").unwrap();
-        self.width = grid_image.width();
-        self.height = grid_image.height();
+        self.width = grid_image.width() as i32;
+        self.height = grid_image.height() as i32;
         self.tiles
             .resize((grid_image.width() * grid_image.height()) as usize, 0);
-        self.tile_size = HEIGHT / self.height as i32;
+        self.tile_size = HEIGHT / self.height;
 
         grid_image.pixels().for_each(|pixel| {
             let x = pixel.0 as usize;
